# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Campaign tool API Documentation",
        "description": "",
        "contact": {
            "email": "hello@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api/v1",
            "description": "V1"
        }
    ],
    "paths": {
...
        "/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Authenticate",
                "description": "Authenticates the user and returns the Bearer token",
                "operationId": "login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            },
                            "example": {
                                "email": "hello@example.com",
                                "password": "password"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "$ref": "#/components/responses/204"
                    },
                    "422": {
                        "$ref": "#/components/responses/422"
                    },
                    "500": {
                        "$ref": "#/components/responses/500"
                    },
                    "503": {
                        "$ref": "#/components/responses/503"
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Unauthenticate",
                "description": "Unauthenticate the user",
                "operationId": "logout",
                "responses": {
                    "204": {
                        "$ref": "#/components/responses/204"
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    },
                    "500": {
                        "$ref": "#/components/responses/500"
                    },
                    "503": {
                        "$ref": "#/components/responses/503"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "responses": {
            "500": {
                "description": "Server Error",
                "content": {
                    "application/json": {
                        "schema": {},
                        "example": {
                            "message": "Server Error."
                        }
                    }
                }
            },
            "503": {
                "description": "Service Unavailable - Maintenance Mode",
                "content": {
                    "application/json": {
                        "schema": {},
                        "example": {
                            "message": "Service Unavailable."
                        }
                    }
                }
            },
            "401": {
                "description": "Unauthenticated",
                "content": {
                    "application/json": {
                        "schema": {},
                        "example": {
                            "message": "Unauthenticated."
                        }
                    }
                }
            },
...
            "422": {
                "description": "Unprocessable Content",
                "content": {
                    "application/json": {
                        "schema": {},
                        "example": {
                            "message": "The phone must be a number. (and 1 more error)",
                            "errors": {
                                "phone": [
                                    "The phone must be a number.",
                                    "The phone must be between 5 and 16 digits."
                                ]
                            }
                        }
                    }
                }
            },
...
            "204": {
                "description": "No Content",
                "content": {
                    "application/json": {
                        "schema": {},
                        "example": []
                    }
                }
            },
...
        },
...
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}